+++ sacred/observers/base.py
@@ -20,7 +20,7 @@ class RunObserver(object):
                       meta_info, _id):
         pass
 
-    def heartbeat_event(self, info, captured_out, beat_time):
+    def heartbeat_event(self, info, captured_out, beat_time, result):
         pass
 
     def completed_event(self, stop_time, result):