+++ sacred/observers/mongo.py
@@ -128,10 +128,11 @@ class MongoObserver(RunObserver):
         self.insert()
         return self.run_entry['_id']
 
-    def heartbeat_event(self, info, captured_out, beat_time):
+    def heartbeat_event(self, info, captured_out, beat_time, result):
         self.run_entry['info'] = flatten(info)
         self.run_entry['captured_out'] = captured_out
         self.run_entry['heartbeat'] = beat_time
+        self.run_entry['result'] = result
         self.save()
 
     def completed_event(self, stop_time, result):