+++ tests/test_observers/test_observers.py
@@ -9,7 +9,7 @@ def test_run_observer():
     # basically to silence coverage
     r = RunObserver()
     assert r.started_event({}, 'run', {}, datetime.utcnow(), {}, 'comment', None) is None
-    assert r.heartbeat_event({}, '', datetime.utcnow()) is None
+    assert r.heartbeat_event({}, '', datetime.utcnow(), 'result') is None
     assert r.completed_event(datetime.utcnow(), 123) is None
     assert r.interrupted_event(datetime.utcnow(), "INTERRUPTED") is None
     assert r.failed_event(datetime.utcnow(), 'trace') is None